feat(auth): refactor auth module into service-oriented architecture
refactor(auth): split auth service into register/login/password/profile/token services
feat(auth): add jwt user lookup service and repository layer
feat(auth): implement login failure tracking and lockout policy
feat(email): add password reset notification template
docs: update architecture and directory structure documentation
chore: update .gitignore and package.json scripts
style: format code with updated prettier config
