+++ .gitignore
@@ -27,6 +27,7 @@ lerna-debug.log*
 *.launch
 .settings/
 *.sublime-workspace
+.qoder
 
 # IDE - VSCode
 .vscode/*
