+++ docs/API_DOCUMENTATION.md
@@ -290,5 +290,6 @@ API调用受速率限制保护：
 ## 版本历史
 
 ### v1.0.0 (最新)
+
 - 初始版本
 - 提供基础认证和会议API
