--- docs/INDEX.md
@@ -1,28 +1,0 @@
-# 文档索引（快捷导航）
-
-更完整的统一入口请查看：`README.md`（本目录）。
-
-常用直达：
-
-- 核心集成
-  - 腾讯会议：`TENCENT_MEETING_INTEGRATION.md`｜`TENCENT_MEETING_WEBHOOK.md`
-  - 飞书（Bitable）：`LARK_INTEGRATION.md`｜`LARK_INTEGRATION_SUMMARY.md`
-
-- 数据与存储
-  - Prisma：`PRISMA_SETUP.md`
-  - 录制文件表：`RECORDING_FILE_TABLE.md`
-
-- 飞书专题
-  - 批量操作：`LARK_BITABLE_BATCH_OPERATIONS.md`
-  - Upsert：`LARK_BITABLE_UPSERT_OPERATIONS.md`
-  - 测试：`LARK_BITABLE_TESTING_GUIDE.md`、`BITABLE_SERVICE_TESTING.md`
-
-- 通信与用户
-  - 邮件：`EMAIL_API.md`
-  - 短信：`ALIYUN_SMS_SETUP.md`
-  - 注册流程：`REGISTRATION_FLOW.md`
-
-- 规范与运维
-  - 项目规范：`PROJECT_GUIDELINES.md`
-  - API 文档：`API_DOCUMENTATION.md`
-  - 部署：`DEPLOYMENT.md`
