+++ docs/README.md
@@ -3,18 +3,21 @@
 本页作为文档统一入口与导航，覆盖集成指南、开发规范、测试与部署等主题。
 
 ## 🚀 快速开始
+
 - 项目总览与运行：根目录 `README.md`
 - 架构与模块：`ARCHITECTURE.md`
 - API 概览与示例：`API_DOCUMENTATION.md`，Swagger：`/api`
 
 ## 🔧 集成与配置
+
 - 腾讯会议：`TENCENT_MEETING_INTEGRATION.md`、`TENCENT_MEETING_WEBHOOK.md`
-- 飞书多维表格（Bitable）：`LARK_INTEGRATION.md`、`LARK_INTEGRATION_SUMMARY.md`
+- 飞书集成：`LARK_INTEGRATION.md`、`LARK_INTEGRATION_SUMMARY.md`、`LARK_WEBHOOK_INTEGRATION.md`
 - 邮件服务：`EMAIL_API.md`
 - 阿里云短信：`ALIYUN_SMS_SETUP.md`
 - Prisma 数据库：`PRISMA_SETUP.md`
 
 ## 📊 飞书多维表格专题
+
 - 录制文件记录表：`RECORDING_FILE_TABLE.md`
 - 批量操作：`LARK_BITABLE_BATCH_OPERATIONS.md`
 - Upsert 操作：`LARK_BITABLE_UPSERT_OPERATIONS.md`
@@ -22,24 +25,54 @@
 - 服务集成测试：`BITABLE_SERVICE_TESTING.md`
 
 ## 👥 用户与认证
+
 - 注册流程：`REGISTRATION_FLOW.md`
 
 ## 🧭 项目规范与协作
+
 - 项目协作与规范（Monorepo 指南）：`PROJECT_GUIDELINES.md`
   - 项目结构、命令、代码风格
   - 测试约定与覆盖率
   - 提交/PR 规范，安全与配置建议
 
 ## 🧪 测试与质量
+
 - 测试分层与运行：[test/README.md](../test/README.md)
 - 常用脚本：根 `README.md` 与 `package.json` scripts
 
 ## 🚀 部署与运维
+
 - 部署与环境：`DEPLOYMENT.md`
 
-## 📎 目录导航
-- 更细的索引与分组：`INDEX.md`
+## ⚡ 快捷导航
+
+常用直达：
+
+- **核心集成**
+  - 腾讯会议：`TENCENT_MEETING_INTEGRATION.md`｜`TENCENT_MEETING_WEBHOOK.md`
+  - 飞书：`LARK_INTEGRATION.md`｜`LARK_INTEGRATION_SUMMARY.md`｜`LARK_WEBHOOK_INTEGRATION.md`
+
+- **数据与存储**
+  - Prisma：`PRISMA_SETUP.md`
+  - 录制文件表：`RECORDING_FILE_TABLE.md`
+
+- **飞书专题**
+  - 批量操作：`LARK_BITABLE_BATCH_OPERATIONS.md`
+  - Upsert：`LARK_BITABLE_UPSERT_OPERATIONS.md`
+  - 测试：`LARK_BITABLE_TESTING_GUIDE.md`｜`BITABLE_SERVICE_TESTING.md`
+
+- **通信与用户**
+  - 邮件：`EMAIL_API.md`
+  - 短信：`ALIYUN_SMS_SETUP.md`
+  - 注册流程：`REGISTRATION_FLOW.md`
+
+- **规范与运维**
+  - 项目规范：`PROJECT_GUIDELINES.md`
+  - API文档：`API_DOCUMENTATION.md`
+  - 部署：`DEPLOYMENT.md`
+  - 目录结构：`DIRECTORY_STRUCTURE.md`
 
 ## 📞 技术支持
-- 优先参考对应文档“故障排除/注意事项”
+
+- 优先参考对应文档"故障排除/注意事项"
 - 如仍有问题，请在任务/Issue 中描述现象与日志
