+++ libs/common/email-templates/index.spec.ts
@@ -1,6 +1,7 @@
 import {
   buildVerificationEmail,
   buildWelcomeEmail,
+  buildPasswordResetNotificationEmail,
 } from '@libs/common/email-templates';
 
 describe('email-templates', () => {
@@ -31,6 +32,23 @@ describe('email-templates', () => {
       const { subject, html } = buildWelcomeEmail(username);
       expect(subject).toBe('欢迎加入 LuLab！');
       expect(html).toContain('LuLab');
+      expect(html).toContain(username);
+      expect(html).toContain('<div');
+      expect(html).toContain('</div>');
+    });
+  });
+
+  describe('buildPasswordResetNotificationEmail', () => {
+    it('builds subject and html including name and time', () => {
+      const username = 'Bob';
+      const fixedDate = new Date('2024-01-02T03:04:05Z');
+      const { subject, html } = buildPasswordResetNotificationEmail(
+        username,
+        fixedDate,
+      );
+      expect(subject).toBe('LuLab 密码重置通知');
+      expect(html).toContain('LuLab');
+      expect(html).toContain('密码重置通知');
       expect(html).toContain(username);
       expect(html).toContain('<div');
       expect(html).toContain('</div>');
