+++ libs/security/jwt/jwt.strategy.ts
@@ -1,15 +1,16 @@
-import { Injectable, UnauthorizedException } from '@nestjs/common';
+import { Injectable, Inject, UnauthorizedException } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { ExtractJwt, Strategy } from 'passport-jwt';
-import { PrismaService } from '@/prisma.service';
 import { ConfigService } from '@nestjs/config';
 import type { JwtPayload, AuthenticatedUser } from '../types';
+import { JWT_USER_LOOKUP, type JwtUserLookup } from '../types';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
   constructor(
-    private readonly prisma: PrismaService,
     private readonly configService: ConfigService,
+    @Inject(JWT_USER_LOOKUP)
+    private readonly userLookup: JwtUserLookup,
   ) {
     super({
       jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
@@ -19,32 +20,12 @@ export class JwtStrategy extends PassportStrategy(Strategy) {
   }
 
   async validate(payload: JwtPayload): Promise<AuthenticatedUser> {
-    const user = await this.prisma.user.findUnique({
-      where: { id: payload.sub },
-      include: {
-        profile: true,
-      },
-    });
-
-    if (!user) {
+    const authUser = await this.userLookup.getAuthenticatedUserById(
+      payload.sub,
+    );
+    if (!authUser) {
       throw new UnauthorizedException('用户不存在');
     }
-
-    if (!user.active) {
-      throw new UnauthorizedException('账户已被禁用');
-    }
-
-    if (user.deletedAt) {
-      throw new UnauthorizedException('账户已被删除');
-    }
-
-    return {
-      id: user.id,
-      username: user.username,
-      email: user.email || '',
-      phone: user.phone,
-      countryCode: user.countryCode,
-      profile: user.profile,
-    };
+    return authUser;
   }
 }
