+++ src/auth/enums/login-type.enum.ts
@@ -1,0 +1,8 @@
+export enum LoginType {
+  USERNAME_PASSWORD = 'USERNAME_PASSWORD',
+  EMAIL_PASSWORD = 'EMAIL_PASSWORD',
+  EMAIL_CODE = 'EMAIL_CODE',
+  PHONE_PASSWORD = 'PHONE_PASSWORD',
+  PHONE_CODE = 'PHONE_CODE',
+  PASSWORD_RESET = 'PASSWORD_RESET',
+}
