+++ src/auth/enums/verification-type.enum.ts
@@ -1,0 +1,6 @@
+export enum VerificationType {
+  EMAIL_VERIFICATION = 'EMAIL_VERIFICATION',
+  PHONE_VERIFICATION = 'PHONE_VERIFICATION',
+  PASSWORD_RESET = 'PASSWORD_RESET',
+  LOGIN_VERIFICATION = 'LOGIN_VERIFICATION',
+}
