+++ src/auth/repositories/user.repository.ts
@@ -1,10 +1,9 @@
 import { Injectable } from '@nestjs/common';
 import { PrismaService } from '@/prisma.service';
 import type { User, UserProfile } from '@prisma/client';
-import type { LoginType } from '../types';
 
 @Injectable()
-export class AuthRepository {
+export class UserRepository {
   constructor(private readonly prisma: PrismaService) {}
 
   // User queries
@@ -145,45 +144,5 @@ export class AuthRepository {
       },
       include: { profile: true },
     });
-  }
-
-  // Login logs
-  countLoginFailuresByTargetSince(
-    target: string,
-    since: Date,
-  ): Promise<number> {
-    return this.prisma.loginLog.count({
-      where: { target, success: false, createdAt: { gte: since } },
-    });
-  }
-
-  countLoginFailuresByIpSince(ip: string, since: Date): Promise<number> {
-    return this.prisma.loginLog.count({
-      where: { ip, success: false, createdAt: { gte: since } },
-    });
-  }
-
-  createLoginLog(data: {
-    userId: string | null;
-    target: string;
-    loginType: LoginType;
-    success: boolean;
-    ip: string;
-    userAgent?: string;
-    failReason?: string;
-  }): Promise<void> {
-    return this.prisma.loginLog
-      .create({
-        data: {
-          userId: data.userId,
-          target: data.target,
-          loginType: data.loginType,
-          success: data.success,
-          ip: data.ip,
-          userAgent: data.userAgent,
-          failReason: data.failReason,
-        },
-      })
-      .then(() => undefined);
   }
 }
