+++ src/verification/dto/verify-code.dto.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsEnum, IsString, MinLength } from 'class-validator';
-import { CodeType } from '@/verification/enums/code-type.enum';
+import { CodeType } from '@/verification/enums';
 
 export class VerifyCodeDto {
   @ApiProperty({ description: '目标邮箱或手机号' })
